tsconfig and index file
