+++ libs/remix-lib/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+      "types": ["node", "tape"],
+      "esModuleInterop": true
+    },
+    "include": ["**/*.ts"]
+  }
